Support Pebble -strict/POST-as-GET
